// frontend/src/types/types.ts
export interface PatientCreate {
  /**
   * First name
   * @minLength 1
   */
  first_name: string;
  /**
   * Last name
   * @minLength 1
   */
  last_name: string;
  /**
   * Dob
   * @format date
   */
  dob: string;
  /** Sex */
  sex: "M" | "F" | "O";
  /**
   * Phone
   * @minLength 1
   */
  phone: string;
  /**
   * Email
   * @format email
   * @minLength 1
   */
  email: string;
  /**
   * Hospital id
   * @format uuid
   */
  // hospital_id: string;
}